Conflicts:
	Dedicated Server/Bin/x86/WorldObjects.cfg
	Dedicated Server/Code/Server/EngineFiles/Server/Host.h
	Source/GameFiles/ClientSide/Client.cpp

+++ Dedicated Server/Code/Server/EngineFiles/Server/ObjectManager.cpp
@@ -274,7 +274,7 @@ bool ObjectManager::InterpCommand(char* command, char* key, WeaponObject* wp)
 
 		sscanf_s(key, "%[^','], %[^','], %[^','], %s", a, sizeof(a), b, sizeof(b), c, sizeof(c));
 
-		Vector3 scale(atof(a), atof(b), atof(c));
+		Vector3 scale((float)atof(a), (float)atof(b), (float)atof(c));
 		wp->SetScale(scale);
 	}
 
@@ -330,7 +330,7 @@ bool ObjectManager::InterpCommand(char* command, char* key, FoodObject* fd)
 
 		sscanf_s(key, "%[^','], %[^','], %[^','], %s", a, sizeof(a), b, sizeof(b), c, sizeof(c));
 
-		Vector3 scale(atof(a), atof(b), atof(c));
+		Vector3 scale((float)atof(a), (float)atof(b), (float)atof(c));
 		fd->SetScale(scale);
 	}
 
@@ -390,7 +390,7 @@ bool ObjectManager::InterpCommand(char* command, char* key, ContainerObject* ct)
 
 		sscanf_s(key, "%[^','], %[^','], %[^','], %s", a, sizeof(a), b, sizeof(b), c, sizeof(c));
 
-		Vector3 scale(atof(a), atof(b), atof(c));
+		Vector3 scale((float)atof(a), (float)atof(b), (float)atof(c));
 		ct->SetScale(scale);
 	}
 
@@ -450,7 +450,7 @@ bool ObjectManager::InterpCommand(char* command, char* key, MaterialObject* ma)
 
 		sscanf_s(key, "%[^','], %[^','], %[^','], %s", a, sizeof(a), b, sizeof(b), c, sizeof(c));
 
-		Vector3 scale(atof(a), atof(b), atof(c));
+		Vector3 scale((float)atof(a), (float)atof(b), (float)atof(c));
 		ma->SetScale(scale);
 	}
 
@@ -510,7 +510,7 @@ bool ObjectManager::InterpCommand(char* command, char* key, StaticProjectileObje
 
 		sscanf_s(key, "%[^','], %[^','], %[^','], %s", a, sizeof(a), b, sizeof(b), c, sizeof(c));
 
-		Vector3 scale(atof(a), atof(b), atof(c));
+		Vector3 scale((float)atof(a), (float)atof(b), (float)atof(c));
 		pt->SetScale(scale);
 	}
 	else if (strcmp(command, MODEL_LENGTH.c_str()) == 0)
